moved FileDropdown component to shared ui folder
